[P44] Cave Functionality Update
